`launch` v2: split launchState into new file
